Changed so non-user owners cannot delete playlist

+++ src/main/java/com/makersacademy/toon_together/controller/PlaylistController.java
@@ -1,6 +1,7 @@
 package com.makersacademy.toon_together.controller;
 
 import com.makersacademy.toon_together.model.Playlist;
+import com.makersacademy.toon_together.model.User;
 import com.makersacademy.toon_together.repository.AuthoritiesRepository;
 import com.makersacademy.toon_together.repository.PlaylistRepository;
 import com.makersacademy.toon_together.repository.UserRepository;
@@ -31,17 +32,20 @@ public class PlaylistController {
 
     @PostMapping("/playlists")
     public RedirectView create(@ModelAttribute Playlist playlist, Authentication auth) {
-        playlist.setUser(userRepository.findByUsername(auth.getName()));
+        playlist.setOwner(userRepository.findByUsername(auth.getName()));
         playlistRepository.save(playlist);
         return new RedirectView("/playlists");
 
     }
 
     @DeleteMapping("/playlists")
-    public RedirectView delete(@RequestParam int id) {
-        System.out.println("hello");
-        System.out.println(id);
-        playlistRepository.deleteById(id);
+    public RedirectView delete(@RequestParam int id, Authentication auth) {
+        User user = userRepository.findByUsername(auth.getName());
+        Playlist playlist = playlistRepository.findById(id);
+        if (user.getId() == playlist.getOwner().getId()) {
+            playlistRepository.deleteById(id);
+        }
+//        maybe want to add some sort of pop-up saying you can delete a playlist that isn't yours
         return new RedirectView("/playlists");
     }
 }
